Add Bradford assay to EnzymeKinetics.Rmd
Because we used the Bradford assay rather than
absorbance at 280 to determine the concentration
of the enzyme, I've modified the block to
calculate the concentration from the calibration
curve. The data from the second week's Bradford
assay is included, but the concentration from the first week is an estimate because it appeared to
be outside the range of standards.